Added SQL Samples for Java

+++ azure-mgmt-compute/src/main/java/com/microsoft/azure/management/compute/implementation/ComputeUsageImpl.java
@@ -27,12 +27,12 @@ class ComputeUsageImpl extends WrapperImpl<UsageInner> implements ComputeUsage {
 
     @Override
     public int currentValue() {
-        return  inner().currentValue();
+        return inner().currentValue();
     }
 
     @Override
-    public int limit() {
-        return (int) inner().limit();
+    public long limit() {
+        return inner().limit();
     }
 
     @Override
