+++ azure-mgmt-trafficmanager/src/main/java/com/microsoft/azure/management/trafficmanager/TrafficManagerNestedProfileEndpoint.java
@@ -21,7 +21,7 @@ public interface TrafficManagerNestedProfileEndpoint extends TrafficManagerEndpo
     /**
      * @return the number of child endpoints to be online to consider nested profile as healthy
      */
-    int minimumChildEndpointCount();
+    long minimumChildEndpointCount();
 
     /**
      * @return the location of the traffic that the endpoint handles
